fix uint32 bug

+++ internal/transform/schema.go
@@ -60,7 +60,7 @@ type TransactionOutput struct {
 	FeeAccount                           string         `json:"fee_account,omitempty"`
 	FeeAccountMuxed                      string         `json:"fee_account_muxed,omitempty"`
 	InnerTransactionHash                 string         `json:"inner_transaction_hash,omitempty"`
-	NewMaxFee                            uint32         `json:"new_max_fee,omitempty"`
+	NewMaxFee                            int64          `json:"new_max_fee,omitempty"`
 	LedgerBounds                         string         `json:"ledger_bounds"`
 	MinAccountSequence                   null.Int       `json:"min_account_sequence"`
 	MinAccountSequenceAge                null.Int       `json:"min_account_sequence_age"`
